# backend-api/api/service/openapi.yaml
openapi: 3.0.3
info:
  title: WAS API
  version: "1.0.0"
...
paths:
...
  /api/v1/traffic-profiles:
    get:
      operationId: ListTrafficProfiles
      tags:
        - traffic-profiles
      summary: List traffic profiles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrafficProfile"
    post:
      operationId: CreateTrafficProfile
      tags:
        - traffic-profiles
      summary: Create traffic profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficProfileCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficProfile"
  /api/v1/traffic-profiles/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetTrafficProfile
      tags:
        - traffic-profiles
      summary: Get traffic profile by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficProfile"
        "404":
          description: Not found
    put:
      operationId: UpdateTrafficProfile
      tags:
        - traffic-profiles
      summary: Update traffic profile by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficProfile"
        "404":
          description: Not found
    delete:
      operationId: DeleteTrafficProfile
      tags:
        - traffic-profiles
      summary: Delete traffic profile by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
  /api/v1/traffic-rules:
    get:
      operationId: ListTrafficRules
      tags:
        - traffic-rules
      summary: List traffic rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrafficRule"
    post:
      operationId: CreateTrafficRule
      tags:
        - traffic-rules
      summary: Create traffic rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficRuleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficRule"
  /api/v1/traffic-rules/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetTrafficRule
      tags:
        - traffic-rules
      summary: Get traffic rule by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficRule"
        "404":
          description: Not found
    put:
      operationId: UpdateTrafficRule
      tags:
        - traffic-rules
      summary: Update traffic rule by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficRuleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficRule"
        "404":
          description: Not found
    delete:
      operationId: DeleteTrafficRule
      tags:
        - traffic-rules
      summary: Delete traffic rule by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
...
components:
  schemas:
...
    TrafficProfile:
      type: object
      required:
        - id
        - name
        - is_enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TrafficProfileCreate:
      type: object
      required:
        - name
        - is_enabled
      properties:
        name:
          type: string
        description:
          type: string
        is_enabled:
          type: boolean
    TrafficProfileUpdate:
      type: object
      required:
        - name
        - is_enabled
      properties:
        name:
          type: string
        description:
          type: string
        is_enabled:
          type: boolean
    TrafficRule:
      type: object
      required:
        - id
        - profile_id
        - name
        - priority
        - dry_run
        - match_all
        - requests_limit
        - period_seconds
        - is_enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        dry_run:
          type: boolean
        match_all:
          type: boolean
        requests_limit:
          type: integer
          format: int32
        period_seconds:
          type: integer
          format: int32
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        is_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TrafficRuleCreate:
      type: object
      required:
        - profile_id
        - name
        - priority
        - dry_run
        - match_all
        - requests_limit
        - period_seconds
        - is_enabled
      properties:
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        dry_run:
          type: boolean
        match_all:
          type: boolean
        requests_limit:
          type: integer
          format: int32
        period_seconds:
          type: integer
          format: int32
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        is_enabled:
          type: boolean
    TrafficRuleUpdate:
      type: object
      required:
        - profile_id
        - name
        - priority
        - dry_run
        - match_all
        - requests_limit
        - period_seconds
        - is_enabled
      properties:
        profile_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          format: int32
        dry_run:
          type: boolean
        match_all:
          type: boolean
        requests_limit:
          type: integer
          format: int32
        period_seconds:
          type: integer
          format: int32
        conditions:
          $ref: "#/components/schemas/SecurityRuleConditions"
          nullable: true
        is_enabled:
          type: boolean
    SecurityHeaderCondition:
      type: object
      required:
        - name
        - value_regex
      properties:
        name:
          type: string
        value_regex:
          type: array
          items:
            type: string
    SecurityRuleConditions:
      type: object
      description: |
        Условия для security rules. Блоки разных типов объединяются по AND.
        Внутри одного блока значения объединяются по OR.
      properties:
        source_ip_cidr:
          type: array
          items:
            type: string
          description: CIDR диапазоны клиентских IP
        uri_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для HTTP URI
        host_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для host
        method_regex:
          type: array
          items:
            type: string
          description: Регулярные выражения для HTTP метода
        headers:
          type: array
          items:
            $ref: "#/components/schemas/SecurityHeaderCondition"
          description: Условия по имени заголовка и regexp значения
...